requirements.txt
sentence-transformers
faiss-cpu
rich
numpy
requests